requirements.txt
streamlit
pandas
plotly
groq
python-dotenv
openpyxl
scikit-learn
statsmodels
